Fix and cleanup Makefile
Some regressions were introduced in 2b6e30c, this fixes those.

Additionally, we switch from a phony build target to actually depending
on changes to markdown to recompile the site content.

Also, we stop removing vendor/bundle in the clean target, since rubygems
are not managed by the makefile anymore so it's inappropriate