Add Travis CI
This moves the CI to travis CI, except for the deployment bits, which
are a bit awkward to get set up for CD in a public repo, as there are
many security concerns surrounding the keys and who can access them.

A new deployment model will be considered shortly. For now, manual
deployments will suffice.