Simplify build system by dropping Sass
This isn't really a necessary thing anymore, especially for a site as
straightforward as this one. We can survive on plain CSS just fine.

In a similar "simplifying" vein, we also rip out the custom image and
css makefile steps, and just have them be carried verbatim by `zod`
which does this out of box.