Switch to a pure Makefile build system
Instead of the semi-working system we had previously, where the entire
site was generated every time one file changed, and the Makefile really
was just a script runner, we're instead now using the Makefile properly
and rendering everything ourselves using a small Ruby script that runs
ERB over our template. We already were using Ruby to generate the
markdown, so this was only a few extra lines to replace an entire static
site generator.

make is truly a great and powerful tool.