Convert site to Hugo
The current Makefile- and Ruby-driven build system is adequate, but is
somewhat inflexible in that we have to maintain it ourselves, lacks many
advanced features, and might scare off others from trying to build or
contribute to the site.

Hugo is a very fast, popular, and feature-rich static stie generator
that we can fairly easily switch to, since all our content is helpfully
already in markdown.

We should convert the entire site to Hugo.